+++ weld/codegen/llvm/resources/vector/serialize_with_pointers.ll
