+++ weld/optimizer/transforms/loop_fusion.rs
@@ -14,6 +14,9 @@ use super::inliner::inline_apply;
 
 use util::SymbolGenerator;
 
+#[cfg(test)]
+use tests::*;
+
 /// Fuses for loops over the same vector in a zip into a single for loop which produces a vector of
 /// structs directly.
 ///
@@ -332,3 +335,115 @@ fn match_types(root_ty: &Type, expr: &mut Expr) {
         _ => {}
     };
 }
+
+#[test]
+fn simple_horizontal_loop_fusion() {
+    // Two loops.
+    let mut e1 = typed_expression("for(zip(
+            result(for([1,2,3], appender, |b,i,e| merge(b, e+1))),
+            result(for([1,2,3], appender,|b2,i2,e2| merge(b2,e2+1)))
+        ), appender, |b,i,e| merge(b, e.$0+1))");
+    fuse_loops_horizontal(&mut e1);
+    let e2 = typed_expression("for(result(for([1,2,3], appender, |b,i,e| merge(b, {e+1,e+1}))), \
+                               appender, |b,i,e| merge(b, e.$0+1))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Three loops.
+    let mut e1 = typed_expression("for(zip(
+            result(for([1,2,3], appender, |b,i,e| merge(b, e+1))),
+            result(for([1,2,3], appender,|b2,i2,e2| merge(b2,e2+2))),
+            result(for([1,2,3], appender,|b3,i3,e3| merge(b3,e3+3)))
+        ), appender, |b,i,e| merge(b, e.$0+1))");
+    fuse_loops_horizontal(&mut e1);
+    let e2 = typed_expression("for(result(for([1,2,3], appender, |b,i,e| merge(b, \
+                               {e+1,e+2,e+3}))), appender, |b,i,e| merge(b, e.$0+1))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Iters in inner loop
+    let mut e1 = typed_expression("for(zip(
+            result(for(iter([1,2,3], 0L, 2L, 1L), appender, |b,i,e| merge(b, e+1))),
+            result(for(iter([1,2,3], 0L, 2L, 1L), appender, |b,i,e| merge(b, e+2)))
+        ), appender, |b,i,e| merge(b, e.$0+1))");
+    fuse_loops_horizontal(&mut e1);
+    let e2 = typed_expression("for(result(for(iter([1,2,3], 0L, 2L, 1L), appender, |b,i,e| \
+                               merge(b, {e+1,e+2}))), appender, |b,i,e| merge(b, e.$0+1))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Iters in outer loop.
+    let mut e1 = typed_expression("for(zip(
+            iter(result(for([1,2,3], appender, |b,i,e| merge(b, e+1))), 0L, 2L, 1L),
+            iter(result(for([1,2,3], appender, |b,i,e| merge(b, e+2))), 0L, 2L, 1L)
+        ), appender, |b,i,e| merge(b, e.$0+1))");
+    fuse_loops_horizontal(&mut e1);
+    let e2 = typed_expression("for(iter(result(for([1,2,3], appender, |b,i,e| merge(b, \
+                               {e+1,e+2}))), 0L, 2L, 1L), appender, |b,i,e| merge(b, e.$0+1))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Two loops with different vectors; should fail.
+    let mut e1 = typed_expression("for(zip(
+            result(for([1,2,3], appender, |b,i,e| merge(b, e+1))),
+            result(for([1,2,4], appender,|b2,i2,e2| merge(b2,e2+1)))
+        ), appender, |b,i,e| merge(b, e.$0+1))");
+    let e2 = e1.clone();
+    fuse_loops_horizontal(&mut e1);
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+}
+
+#[test]
+fn simple_vertical_loop_fusion() {
+    // Two loops.
+    let mut e1 = typed_expression("for(result(for([1,2,3], appender, |b,i,e| merge(b,e+2))), \
+                                   appender, |b,h,f| merge(b, f+1))");
+    fuse_loops_vertical(&mut e1);
+    let e2 = typed_expression("for([1,2,3], appender, |b,i,e| merge(b, (e+2)+1))");
+    println!("{}", print_expr_without_indent(&e1));
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Three loops.
+    let mut e1 = typed_expression("for(result(for(result(for([1,2,3], appender, |b,i,e| \
+                                   merge(b,e+3))), appender, |b,i,e| merge(b,e+2))), appender, \
+                                   |b,h,f| merge(b, f+1))");
+    fuse_loops_vertical(&mut e1);
+    let e2 = typed_expression("for([1,2,3], appender, |b,i,e| merge(b, (((e+3)+2)+1)))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Merges in other positions, replace builder identifiers.
+    let mut e1 = typed_expression("for(result(for([1,2,3], appender, |b,i,e| if(e>5, \
+                                   merge(b,e+2), b))), appender, |b,h,f| merge(b, f+1))");
+    fuse_loops_vertical(&mut e1);
+    let e2 = typed_expression("for([1,2,3], appender, |b,i,e| if(e>5, merge(b, (e+2)+1), b))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Make sure correct builder is chosen.
+    let mut e1 = typed_expression("for(result(for([1,2,3], appender[i32], |b,i,e| \
+                                   merge(b,e+2))), appender[f64], |b,h,f| merge(b, 1.0))");
+    fuse_loops_vertical(&mut e1);
+    let e2 = typed_expression("for([1,2,3], appender[f64], |b,i,e| merge(b, 1.0))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Multiple inner loops.
+    let mut e1 = typed_expression("for(result(for(zip([1,2,3],[4,5,6]), appender, |b,i,e| \
+                                   merge(b,e.$0+2))), appender, |b,h,f| merge(b, f+1))");
+    fuse_loops_vertical(&mut e1);
+    let e2 = typed_expression("for(zip([1,2,3],[4,5,6]), appender, |b,i,e| merge(b, (e.$0+2)+1))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Iter where inner data consumed fully.
+    let mut e1 = typed_expression("let a = [1,2,3]; for(result(for(iter(a, 0L, len(a), 1L), \
+                                   appender, |b,i,e| merge(b,e+2))), appender, |b,h,f| merge(b, \
+                                   f+1))");
+    fuse_loops_vertical(&mut e1);
+    let e2 = typed_expression("let a = [1,2,3]; for(iter(a,0L,len(a),1L), appender, |b,i,e| \
+                               merge(b, (e+2)+1))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+
+    // Inner data not consumed fully.
+    let mut e1 = typed_expression("for(result(for(iter([1,2,3], 0L, 1L, 1L), appender, |b,i,e| \
+                                   merge(b,e+2))), appender, |b,h,f| merge(b, f+1))");
+    fuse_loops_vertical(&mut e1);
+    // Loop fusion should fail.
+    let e2 = typed_expression("for(result(for(iter([1,2,3], 0L, 1L, 1L), appender, |b,i,e| \
+                               merge(b,e+2))), appender, |b,h,f| merge(b, f+1))");
+    assert!(e1.compare_ignoring_symbols(&e2).unwrap());
+}
+
