+++ weld/syntax/mod.rs
@@ -1,4 +1,8 @@
+//! Syntax of the Weld language.
+//!
+//! This module contains sub-modules that tokenize and parse Weld programs and their macros.
 
 pub mod macro_processor;
 pub mod parser;
+pub mod program;
 pub mod tokenizer;
